// app/utils/type.ts
export type initialPoolInfoType = {
    id: number;
    tokenA: string;
    tokenB: string;
    favourite: boolean;
    arranged: boolean;
    liquidity: number;
    volume24h: number;
    volume7d: number;
    volume30d: number;
    fees24h: number;
    fees7d: number;
    fees30d: number;
    apr24h: number;
    apr7d: number;
    apr30d: number;
  };